+++ src/client/lazy-app/Compress/stages/compress-stage.ts
@@ -1,0 +1,31 @@
+import { encoderMap, EncoderState } from 'client/lazy-app/feature-meta';
+import { assertSignal, ImageMimeTypes } from 'client/lazy-app/util';
+import WorkerBridge from 'client/lazy-app/worker-bridge';
+
+export async function compressImage(
+  signal: AbortSignal,
+  image: ImageData,
+  encodeData: EncoderState,
+  sourceFilename: string,
+  workerBridge: WorkerBridge,
+): Promise<File> {
+  assertSignal(signal);
+
+  const encoder = encoderMap[encodeData.type];
+  const compressedData = await encoder.encode(
+    signal,
+    workerBridge,
+    image,
+    // The type of encodeData.options is enforced via the previous line
+    encodeData.options as any,
+  );
+
+  // This type ensures the image mimetype is consistent with our mimetype sniffer
+  const type: ImageMimeTypes = encoder.meta.mimeType;
+
+  return new File(
+    [compressedData],
+    sourceFilename.replace(/.[^.]*$/, `.${encoder.meta.extension}`),
+    { type },
+  );
+}
